backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
httpie==3.2.2
idna==3.7
Markdown==3.6
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
pillow==10.3.0
psycopg2-binary==2.9.9
Pygments==2.17.2
PySocks==1.7.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.1
sqlparse==0.5.0
urllib3==2.2.1

